feat: add command line args
- parameterise samples input file
- add dry run option
- update docs

+++ finetune/llamafinetune.py
@@ -1,6 +1,7 @@
 import sys
 import os
 import json
+import argparse
 from datetime import datetime
 from dotenv import load_dotenv
 from sagemaker.jumpstart.estimator import JumpStartEstimator
@@ -10,6 +11,31 @@ from utils.utils import load_config
 from utils.aws import upload_file
 
 
+def parse_CLI_args() -> argparse.Namespace:
+    """Parse command line arguments
+
+    Returns:
+        args : Namespace
+            Namespace of passed command line argument inputs
+    """
+    parser = argparse.ArgumentParser()
+
+    parser.add_argument(
+        "-f",
+        "--file",
+        type=str,
+        default="anthropic_batch_job.jsonl.out",
+        help="Name of the AWS Bedrock Anthropic batch inference output file containing sample data",
+    )
+    parser.add_argument(
+        "-d",
+        "--dry_run",
+        action="store_true",
+        help="Whether to simulate calling AWS endpoints",
+    )
+    arguments = parser.parse_args()
+    return arguments
+
 def generate_template_file() -> None:
     with open("template.json", "w") as outfile:
         print(json.dumps({
@@ -17,9 +43,9 @@ def generate_template_file() -> None:
         }), 
         file=outfile)
 
-def generate_train_file() -> None:
+def generate_train_file(samples_input_file) -> None:
     with open("train.jsonl", "w") as outfile:
-        with open("anthropic_batch_job.jsonl.out", "r") as infile:
+        with open(samples_input_file, "r") as infile:
             for line in infile:
                 line = json.loads(line)
                 text = line["modelOutput"]["content"][0]["text"].replace("```json", "").replace("```", "")
@@ -64,35 +90,39 @@ def run_finetune_pipeline(
     estimator.fit({'training': input_path})
 
 if __name__ == "__main__":
+    args = parse_CLI_args()
     config = load_config()
     load_dotenv()
 
     job_id = "finetune/" + datetime.now().strftime("%Y-%m-%d-%H%M")
 
     generate_template_file()
-    upload_file(
-        config["llama"]["region"], 
-        "template.json",
-        os.getenv("BUCKET"),
-        "template.json",
-        job_id + "/input"
-    )
+    if(not args.dry_run):
+        upload_file(
+            config["llama"]["region"], 
+            "template.json",
+            os.getenv("BUCKET"),
+            "template.json",
+            job_id + "/input"
+        )
 
-    generate_train_file()
-    upload_file(
-        config["llama"]["region"],
-        "train.jsonl",
-        os.getenv("BUCKET"),
-        "train.jsonl",
-        job_id + "/input"
-    )
+    generate_train_file(args.file)
+    if(not args.dry_run):
+        upload_file(
+            config["llama"]["region"],
+            "train.jsonl",
+            os.getenv("BUCKET"),
+            "train.jsonl",
+            job_id + "/input"
+        )
 
-    run_finetune_pipeline(
-        config["llama"]["model"],
-        config["llama"]["version"],
-        os.environ["ROLE"],
-        "s3://" + os.getenv("BUCKET") + "/" + job_id + "/input",
-        "s3://" + os.getenv("BUCKET") + "/" + job_id + "/output",
-        os.environ["INSTANCE_TYPE"],
-        config["llama"]["region"]
-    )  
+    if(not args.dry_run):
+        run_finetune_pipeline(
+            config["llama"]["model"],
+            config["llama"]["version"],
+            os.environ["ROLE"],
+            "s3://" + os.getenv("BUCKET") + "/" + job_id + "/input",
+            "s3://" + os.getenv("BUCKET") + "/" + job_id + "/output",
+            os.environ["INSTANCE_TYPE"],
+            config["llama"]["region"]
+        )
